Created new branch
This is the shape buttons riddle.

In this riddle, the Manager script shuffles the given textures and then assigns them to the four ShapeButtons in the scene. A dictionary holds each textures' value and this decides the order in which the buttons need to be pressed.

If you want to test the game, run it and look at the ShapeButtonManager in the inspector. The order in which the names of buttons appear (from top to bottom) is the order they need to be pressed to pass the puzzle.