Create new branch
Created new branch for the wire cutting puzzle.

Do not be alarmed by the empty scene. The WireManager spawns and alligns between 4 and 6 wires with random colors, and later calculates which wire needs to be cut through various condition checks.

If you want to test the game, run it and look at the console. A debug command in the code should told you the number of the wire that needs to be cut. This does not follow array logic, meaning if it says "wire number 2", you cut the 2nd wire, not the 3rd.