Zoom Camera
Created a script with a Zoom In and Zoom Out Coroutines.

Canvas elements have been divided into two groups held by an empty game object, Zoom In Elements and Zoom Out Elements. This is for the purpose of easily switching between which elements need to appear and disappear when the camera is zoomed in or out using the SetActive(bool) method.

The zoom script moves the camera to a target position while changing the orthogaphic camera size to create a zoom in and out effect. Currently the size values are set in the code and are not editable.