Concluded Test
Results:
Mechanically, everything is working as it should.

Bugs:
Sounds don't play in Unity Remote 5, but work when the game is built and run as an APK file. It is likely safe to assume that if sounds play on a computer while testing, they will play on mobile devices as well.