New Gameplay Features & Script Updates
All main puzzle managers now inherit from "Manager" class. This is used for camera focus and easy management for test if all puzzles are solved.

Added new "GameManager" which controls and checks the status of all five puzzles to verify when the game is won, as well as manager the amount of strikes a player has.

Added new "GameUIManager" which controls the appearance and disappearance of the zoom buttons.

The "FocusCamera" method now takes objects of type Manager as a parameter and saves them under the new focusedManager variable of type Manager. This is used for enabling and dsiabling scripts when zooming in and out of them. Additionally, this method now also invokes GameUIManager methods to show and hide zoom buttons.

TimeManager now has a new method "AccelerateTime" which reduces the countdownDelay variable, causing the timer to count down faster. This method also invokes the StrikeDown() method from the GameManager.

Updated all puzzles with aforementioned scripts and methods, except SafeLock puzzle.